aadd

+++ landmark.py
@@ -2,8 +2,9 @@ import time
 
 import cv2
 from picamera2 import Picamera2
-from calibrate_camera_constants import CameraMatrix, DistortionCoefficient, markerHeight
+
 import robot
+from calibrate_camera_constants import CameraMatrix, DistortionCoefficient, markerHeight
 
 arlo = robot.Robot()
 
@@ -146,8 +147,15 @@ while 1 and __name__ == "__main__":
                 2,
             )
             print("[INFO] ArUco marker ID: {}".format(markerID))
-            
-            print(cv2.aruco.estimatePoseSingleMarkers(corners, markerHeight, CameraMatrix(preview_downscale), DistortionCoefficient))
+
+            print(
+                cv2.aruco.estimatePoseSingleMarkers(
+                    corners,
+                    markerHeight,
+                    CameraMatrix(preview_downscale).astype('float32'),
+                    DistortionCoefficient.astype('float32'),
+                )
+            )
 
             # show the output image
     cv2.imshow("Image", image)
